requirements.txt
agate==1.6.1
airflow-dbt==0.3.0
alembic==1.4.3
apache-airflow==1.10.12
apispec==1.3.3
argcomplete==1.12.1
asn1crypto==1.4.0
attrs==19.3.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
Babel==2.8.0
boto3==1.11.17
botocore==1.14.17
cached-property==1.5.2
cachetools==4.1.1
cattrs==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.35
cryptography==2.9.2
dbt==0.18.1
dbt-bigquery==0.18.1
dbt-core==0.18.1
dbt-postgres==0.18.1
dbt-redshift==0.18.1
dbt-snowflake==0.18.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
dnspython==2.0.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
google-api-core==1.16.0
google-auth==1.22.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.6.0
graphviz==0.14.2
gunicorn==20.0.4
hologram==0.0.10
idna==2.9
importlib-metadata==2.0.0
iso8601==0.1.13
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
json-merge-patch==0.2
json-rpc==1.13.0
jsonschema==3.1.1
keyring==21.4.0
lazy-object-proxy==1.5.1
leather==0.3.3
lockfile==0.12.2
Logbook==1.5.3
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.0
minimal-snowplow-tracker==0.0.2
msrest==0.6.19
natsort==7.0.1
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.1.3
parsedatetime==2.6
pendulum==1.4.4
prison==0.1.3
protobuf==3.11.3
psutil==5.7.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.8
Pygments==2.7.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
setproctitle==1.1.10
six==1.15.0
snowflake-connector-python==2.2.10
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.1
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
Werkzeug==0.16.1
WTForms==2.3.3
zipp==3.4.0
zope.deprecation==4.4.0
